#13 Remove code duplication
