+++ pkg/domain/logs.go
@@ -6,8 +6,3 @@ type LogLine struct {
 	Timestamp time.Time
 	Value     string
 }
-
-type LogLine struct {
-	Timestamp time.Time
-	Value     string
-}
